requirements.txt
requests 
pytest 
python-dotenv 
apache-airflow 